refactor: 미사용 필드 제거
