+++ src/main/java/kitchenpos/menu/application/dto/MenuProductDto.java
@@ -1,27 +1,13 @@
 package kitchenpos.menu.application.dto;
 
 public class MenuProductDto {
-    private Long seq;
-    private Long menuId;
     private Long productId;
     private Long quantity;
 
-    public MenuProductDto(final Long seq,
-                          final Long menuId,
-                          final Long productId,
+    public MenuProductDto(final Long productId,
                           final Long quantity) {
-        this.seq = seq;
-        this.menuId = menuId;
         this.productId = productId;
         this.quantity = quantity;
-    }
-
-    public Long getSeq() {
-        return seq;
-    }
-
-    public Long getMenuId() {
-        return menuId;
     }
 
     public Long getProductId() {
