+++ src/main/java/kitchenpos/menu/application/dto/MenuProductResponse.java
@@ -1,0 +1,34 @@
+package kitchenpos.menu.application.dto;
+
+public class MenuProductResponse {
+    private Long seq;
+    private Long menuId;
+    private Long productId;
+    private Long quantity;
+
+    public MenuProductResponse(final Long seq,
+                               final Long menuId,
+                               final Long productId,
+                               final Long quantity) {
+        this.seq = seq;
+        this.menuId = menuId;
+        this.productId = productId;
+        this.quantity = quantity;
+    }
+
+    public Long getSeq() {
+        return seq;
+    }
+
+    public Long getMenuId() {
+        return menuId;
+    }
+
+    public Long getProductId() {
+        return productId;
+    }
+
+    public Long getQuantity() {
+        return quantity;
+    }
+}
